requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
numpy
openai
torch
transformers
pdfplumber
python-docx
python-dotenv
huggingface-hub
accelerate
unstructured
asyncio
aiohttp
aiogram
sqlalchemy
pgvector
psycopg2